Revert "[Refactor] Implement route-level bundle splitting"
