+++ package.json
@@ -58,12 +58,11 @@
     "@babel/plugin-proposal-class-properties": "7.3.4",
     "@babel/plugin-proposal-decorators": "7.3.0",
     "@babel/plugin-proposal-json-strings": "7.2.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "7.2.0",
     "@babel/plugin-syntax-import-meta": "7.0.0",
     "@babel/preset-env": "7.3.4",
     "@babel/preset-react": "7.0.0",
     "@babel/preset-typescript": "7.7.2",
-    "@loadable/babel-plugin": "^5.12.0",
     "@omakase/cli": "0.0.10",
     "@sentry/browser": "4.2.3",
     "@storybook/addon-info": "5.1.10",
@@ -102,6 +101,7 @@
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "24.8.0",
     "babel-loader": "8.0.5",
+    "babel-plugin-dynamic-import-node": "1.2.0",
     "babel-plugin-lodash": "3.3.3",
     "babel-plugin-module-resolver": "3.1.1",
     "babel-plugin-relay": "7.1.0",
@@ -170,8 +170,6 @@
     "@artsy/detect-responsive-traits": "^0.0.5",
     "@artsy/fresnel": "^1.0.13",
     "@artsy/react-html-parser": "^3.0.2",
-    "@loadable/component": "^5.12.0",
-    "@loadable/server": "^5.12.0",
     "@types/luxon": "^1.15.1",
     "autosuggest-highlight": "^3.1.1",
     "cheerio": "^1.0.0-rc.2",
