+++ src/Apps/Artist/Routes/Articles/index.tsx
@@ -30,6 +30,3 @@ export const ArticlesRouteFragmentContainer = createFragmentContainer(
     `,
   }
 )
-
-// Top-level route needs to be exported for bundle splitting in the router
-export default ArticlesRouteFragmentContainer
