+++ src/Apps/Artist/Routes/AuctionResults/index.tsx
@@ -21,6 +21,3 @@ export const AuctionResultsRouteFragmentContainer = createFragmentContainer(
     `,
   }
 )
-
-// Top-level route needs to be exported for bundle splitting in the router
-export default AuctionResultsRouteFragmentContainer
