+++ src/Apps/Artist/Routes/Shows/index.tsx
@@ -76,6 +76,3 @@ export const ShowsRouteFragmentContainer = createFragmentContainer(ShowsRoute, {
     }
   `,
 })
-
-// Top-level route needs to be exported for bundle splitting in the router
-export default ShowsRouteFragmentContainer
