+++ src/Apps/Artist/Routes/Works/index.tsx
@@ -141,6 +141,3 @@ export const WorksRouteFragmentContainer = createFragmentContainer(WorksRoute, {
     }
   `,
 })
-
-// Top-level route needs to be exported for bundle splitting in the router
-export default WorksRouteFragmentContainer
