+++ src/Apps/Artist/routes.tsx
@@ -1,4 +1,3 @@
-import loadable from "@loadable/component"
 import { hasSections as showMarketInsights } from "Apps/Artist/Components/MarketInsights/MarketInsights"
 import {
   ArtworkFilters,
@@ -9,7 +8,14 @@ import { paramsToCamelCase } from "Components/v2/ArtworkFilter/Utils/urlBuilder"
 import { Redirect, RedirectException, RouteConfig } from "found"
 import * as React from "react"
 import { graphql } from "react-relay"
+import { ArtistAppFragmentContainer as ArtistApp } from "./ArtistApp"
 import { hasOverviewContent } from "./Components/NavigationTabs"
+import { ArticlesRouteFragmentContainer as ArticlesRoute } from "./Routes/Articles"
+import { AuctionResultsRouteFragmentContainer as AuctionResultsRoute } from "./Routes/AuctionResults"
+import { CVRouteFragmentContainer as CVRoute } from "./Routes/CV"
+import { OverviewRouteFragmentContainer as OverviewRoute } from "./Routes/Overview"
+import { ShowsRouteFragmentContainer as ShowsRoute } from "./Routes/Shows"
+import { WorksRouteFragmentContainer as WorksRoute } from "./Routes/Works"
 
 graphql`
   fragment routes_Artist on Artist {
@@ -62,7 +68,7 @@ graphql`
 export const routes: RouteConfig[] = [
   {
     path: "/artist/:artistID",
-    getComponent: () => loadable(() => import("./ArtistApp")),
+    Component: ArtistApp,
     query: graphql`
       query routes_ArtistTopLevelQuery($artistID: String!) @raw_response_type {
         artist(id: $artistID) @principalField {
@@ -100,7 +106,7 @@ export const routes: RouteConfig[] = [
     children: [
       {
         path: "/",
-        getComponent: () => loadable(() => import("./Routes/Overview")),
+        Component: OverviewRoute,
         displayNavigationTabs: true,
         query: graphql`
           query routes_OverviewQuery($artistID: String!) @raw_response_type {
@@ -112,7 +118,7 @@ export const routes: RouteConfig[] = [
       },
       {
         path: "cv",
-        getComponent: () => loadable(() => import("./Routes/CV")),
+        Component: CVRoute,
         query: graphql`
           query routes_CVQuery($artistID: String!) {
             viewer {
@@ -123,7 +129,7 @@ export const routes: RouteConfig[] = [
       },
       {
         path: "articles",
-        getComponent: () => loadable(() => import("./Routes/Articles")),
+        Component: ArticlesRoute,
         query: graphql`
           query routes_ArticlesQuery($artistID: String!) {
             artist(id: $artistID) {
@@ -134,7 +140,7 @@ export const routes: RouteConfig[] = [
       },
       {
         path: "shows",
-        getComponent: () => loadable(() => import("./Routes/Shows")),
+        Component: ShowsRoute,
         query: graphql`
           query routes_ShowsQuery($artistID: String!) {
             viewer {
@@ -145,7 +151,7 @@ export const routes: RouteConfig[] = [
       },
       {
         path: "auction-results",
-        getComponent: () => loadable(() => import("./Routes/AuctionResults")),
+        Component: AuctionResultsRoute,
         displayNavigationTabs: true,
         query: graphql`
           query routes_AuctionResultsQuery($artistID: String!) {
@@ -157,7 +163,7 @@ export const routes: RouteConfig[] = [
       },
       {
         path: "works-for-sale",
-        getComponent: () => loadable(() => import("./Routes/Works")),
+        Component: WorksRoute,
         displayNavigationTabs: true,
         query: graphql`
           query routes_WorksQuery(
