+++ src/Apps/Artwork/ArtworkApp.tsx
@@ -309,6 +309,3 @@ export const ArtworkAppFragmentContainer = createFragmentContainer(
     `,
   }
 )
-
-// Top-level route needs to be exported for bundle splitting in the router
-export default ArtworkAppFragmentContainer
