+++ src/Apps/Artwork/routes.tsx
@@ -1,10 +1,15 @@
-import loadable from "@loadable/component"
 import { graphql } from "react-relay"
+import { ArtworkAppFragmentContainer as ArtworkApp } from "./ArtworkApp"
+
+// @ts-ignore
+import { ComponentClass, StatelessComponent } from "react"
+
+// TODO: Investigate better error boundaries for runtime errors
 
 export const routes = [
   {
     path: "/artwork/:artworkID/(confirm-bid)?",
-    getComponent: () => loadable(() => import("./ArtworkApp")),
+    Component: ArtworkApp,
     query: graphql`
       query routes_ArtworkQuery($artworkID: String!) {
         artwork(id: $artworkID) @principalField {
