+++ src/Apps/Search/SearchApp.tsx
@@ -169,6 +169,3 @@ export const SearchAppFragmentContainer = createFragmentContainer(
     `,
   }
 )
-
-// Top-level route needs to be exported for bundle splitting in the router
-export default SearchAppFragmentContainer
