+++ src/Apps/Search/routes.tsx
@@ -1,13 +1,15 @@
-import loadable from "@loadable/component"
 import { RouteConfig } from "found"
 import React from "react"
 import { graphql } from "react-relay"
 
+import { SearchResultsArtistsRouteFragmentContainer as SearchResultsArtistsRoute } from "Apps/Search/Routes/Artists/SearchResultsArtists"
+import { SearchResultsArtworksRoute } from "Apps/Search/Routes/Artworks"
 import { SearchResultsEntityRouteFragmentContainer as SearchResultsEntityRoute } from "Apps/Search/Routes/Entity/SearchResultsEntity"
 
 import { RouteSpinner } from "Artsy/Relay/renderWithLoadProgress"
-import { ArtworkQueryFilter } from "Components/v2/ArtworkFilter/ArtworkQueryFilter"
+import { ArtworkQueryFilter } from "Components/v2/ArtworkFilter"
 import { paramsToCamelCase } from "Components/v2/ArtworkFilter/Utils/urlBuilder"
+import { SearchAppFragmentContainer as SearchApp } from "./SearchApp"
 
 const prepareVariables = (_params, { location }) => {
   return {
@@ -65,7 +67,7 @@ const entityTabs = Object.entries(tabsToEntitiesMap).map(([key, entities]) => {
 export const routes: RouteConfig[] = [
   {
     path: "/search",
-    getComponent: () => loadable(() => import("./SearchApp")),
+    Component: SearchApp,
     query: graphql`
       query routes_SearchResultsTopLevelQuery($term: String!) {
         viewer {
@@ -77,14 +79,13 @@ export const routes: RouteConfig[] = [
     children: [
       {
         path: "/",
-        getComponent: () => loadable(() => import("./Routes/Artworks")),
+        Component: SearchResultsArtworksRoute,
         prepareVariables,
         query: ArtworkQueryFilter,
       },
       {
         path: "artists",
-        getComponent: () =>
-          loadable(() => import("./Routes/Artists/SearchResultsArtists")),
+        Component: SearchResultsArtistsRoute,
         prepareVariables,
         query: graphql`
           query routes_SearchResultsArtistsQuery($term: String!, $page: Int) {
