+++ src/Artsy/Router/buildServerApp.tsx
@@ -1,4 +1,3 @@
-import path from "path"
 import React from "react"
 import ReactDOMServer from "react-dom/server"
 import serialize from "serialize-javascript"
@@ -29,9 +28,6 @@ import { RouterConfig } from "./"
 import { queryStringParsing } from "./Utils/queryStringParsing"
 import { RenderError, RenderPending, RenderReady } from "./Utils/RenderStatus"
 
-import { ChunkExtractor } from "@loadable/server"
-import { getENV } from "Utils/getENV"
-
 interface Resolve {
   bodyHTML?: string
   redirect?: {
@@ -43,7 +39,9 @@ interface Resolve {
   styleTags?: string
 }
 
+// No need to invoke this for each request.
 const MediaStyle = createMediaStyle()
+
 const logger = createLogger("Artsy/Router/buildServerApp.tsx")
 
 export interface ServerRouterConfig extends RouterConfig {
@@ -57,9 +55,7 @@ export function buildServerApp(config: ServerRouterConfig): Promise<Resolve> {
       try {
         const { context = {}, routes = [], url, userAgent } = config
         const user = getUser(context.user)
-        const relayEnvironment =
-          context.relayEnvironment ||
-          createRelaySSREnvironment({ user, userAgent })
+        const relayEnvironment = context.relayEnvironment || createRelaySSREnvironment({ user, userAgent }) // prettier-ignore
         const historyMiddlewares = [
           createQueryMiddleware({
             parse: queryStringParsing,
@@ -90,14 +86,11 @@ export function buildServerApp(config: ServerRouterConfig): Promise<Resolve> {
           resolve({ redirect: farceResults.redirect })
         } else {
           /**
-           * An array that gets passed to `react-head`'s provider that will
-           * collect the header tags that are rendered by the App. `headTags`
-           * is _mutated_ when it's passed to the App. Beware.
-           *
+           * An array that gets passed to `react-head`'s provider that will collect the header
+           * tags that are rendered by the App. `headTags` is _mutated_ when it's passed to the App. Beware.
            **/
           const headTags = [<style type="text/css">{MediaStyle}</style>]
-          const matchingMediaQueries =
-            userAgent && matchingMediaQueriesForUserAgent(userAgent)
+          const matchingMediaQueries = userAgent && matchingMediaQueriesForUserAgent(userAgent) // prettier-ignore
 
           const ServerApp = ({ tags = [] }) => {
             return (
@@ -114,89 +107,14 @@ export function buildServerApp(config: ServerRouterConfig): Promise<Resolve> {
             )
           }
 
-          // Build up script tags to inject into head
-          const scripts = []
-
           const { relayData: _relayData, styleTags, bodyHTML } = await trace(
             "buildServerApp.fetch",
             (async () => {
               const sheet = new ServerStyleSheet()
 
-              // Extract bundle split chunks for SSR
-              let bundleJSX
-              let html
-
-              // FIXME: Remove once EXPERIMENTAL_APP_SHELL A/B test completes
-              if (getENV("EXPERIMENTAL_APP_SHELL")) {
-                /**
-                 * This is the stats file generated by Force's webpack setup, via
-                 * @loadable/webpack-plugin.
-                 *
-                 * @see https://github.com/damassi/force/blob/master/webpack/envs/baseConfig.js#L94
-                 */
-                let statsFile
-                try {
-                  statsFile = path.resolve(
-                    process.cwd(),
-                    "public/assets",
-                    "loadable-stats.json"
-                  )
-                } catch (error) {
-                  console.log(
-                    "[Artsy/Router/buildServerApp.tsx] Error:",
-                    "Missing loadable-stats.json file"
-                  )
-                }
-
-                const extractor = new ChunkExtractor({
-                  statsFile,
-                  entrypoints: [
-                    // FIXME: Remove once EXPERIMENTAL_APP_SHELL A/B test completes
-                    // (rename entrypoint)
-                    "experimental-app-shell",
-                  ],
-                })
-
-                // Wrap component tree in library contexts to extract usage
-                bundleJSX = extractor.collectChunks(
-                  sheet.collectStyles(<ServerApp tags={headTags} />)
-                )
-
-                html = ReactDOMServer.renderToString(bundleJSX)
-
-                // Extract all dynamic `import()` bundle split tags from app
-                const bundleScriptTags = extractor.getScriptTags()
-
-                scripts.push(
-                  bundleScriptTags
-                    .split("\n")
-                    .filter(
-                      script =>
-                        !(
-                          /**
-                           * Since these files are already embedded in our main
-                           * layout file, omit them from the scripts array.
-                           *
-                           * TODO: Don't include these files in the main layout
-                           * and instead relay on dynamically including them here.
-                           * Will require some shuffling in the jade template,
-                           * however.
-                           */
-                          (
-                            script.includes("/assets/runtime-manifest.") ||
-                            script.includes("/assets/common.")
-                          )
-                        )
-                    )
-                    .join("\n")
-                )
-
-                // TODO: Wire up if A/B test is successful
-                // extractor.getLinkTags()
-              } else {
-                bundleJSX = sheet.collectStyles(<ServerApp tags={headTags} />)
-                html = ReactDOMServer.renderToString(bundleJSX)
-              }
+              const html = ReactDOMServer.renderToString(
+                sheet.collectStyles(<ServerApp tags={headTags} />)
+              )
 
               // Get serializable Relay data for rehydration on the client
               const data = await relayEnvironment.relaySSRMiddleware.getCache()
@@ -215,11 +133,13 @@ export function buildServerApp(config: ServerRouterConfig): Promise<Resolve> {
           // Strip response of problematic data structures
           const relayData = cleanRelayData(_relayData)
 
+          // Build up script tags to inject into head
+          const scripts = []
           scripts.push(`
-            <script>
-              var __RELAY_BOOTSTRAP__ = ${serializeRelayData(relayData)};
-            </script>
-          `)
+          <script>
+            var __RELAY_BOOTSTRAP__ = ${serializeRelayData(relayData)};
+          </script>
+        `)
 
           const result = {
             bodyHTML,
